recorder-desktop/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "recorder-desktop"
version = "0.1.0"
description = "Desktop recorder skeleton"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "pywinauto>=0.6.8",
    "pynput>=1.7.6",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["recorder_desktop*"]
